requirements.txt
Flask
gunicorn
psycopg2-binary
python-dotenv
requests
bcrypt
supabase
python-jose[cryptography]  # For JWT handling
gotrue  # For Supabase Auth
postgrest  # For Supabase realtime
realtime  # For Supabase realtime

# 🖥️ PyQt6 for Desktop GUI
PyQt6>=6.4.0

# 🧠 مكتبات الذكاء الاصطناعي والتحليل المتقدم
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pillow>=10.0.0
pyzbar>=0.1.9
qrcode>=7.4.0
